(SIMP-3502) core: Fix Forge publishing
Previously, the travis.yml did not look like our other modules and did
not live in it's own.

The purpose of having simp-core have it's own puppet module is to behave
as a shortcut to install SIMP from the Puppet Forge. It only has
dependencies required to get started.

Full previous discussion can be found here:
https://github.com/simp/simp-core/pull/427

SIMP-3502 #close
SIMP-3881 #close